Update to eslint 9 (#40555)
Eslint 9 brings a new config format, which means this PR also rewrites
all the eslint configs. I took the opportunity to clean them up
(although a good bit of that wound up in earlier PRs), and added code to
fetch the textdomain from composer.json and whether the project uses
React from package.json so we don't need to have so many boilerplate
configs that just set a textdomain or enable React rules.