codeql: Remove unneeded advanced config file (#40674)
GitHub has a "default" config and an "advanced" config. It turns out the
"advanced" config we have here doesn't do anything very advanced, we can
switch to the default config.

See https://github.com/Automattic/jetpack/security/code-scanning/438

When I did that, it disabled the workflow but didn't create a PR to
delete this file for us. Guess we have to do that manually.