Pages: Add quick actions to change the homepage and the posts page (#40699)
In order to remove the Pages screen from Calypso, we need to port some missing features to WP Admin first.

One of these features is the ability to change the homepage and the posts page directly from the pages list. This commit implements it in WP Admin as quick actions displayed in the pages list.